Move Notification to its own file
